+++ docs/CHANGES.md
@@ -4,5 +4,6 @@
 - Updated - Event registration for pet level-ups and shared events - [core.lua, initialization.lua]
 - Updated - Delve Companion level-up detection using chat messages - [core.lua, initialization.lua]
 - Removed - Sound throttling for reputation events - [core.lua]
+- Added - Proccess queue system - [utils.lua]
 - Added - New debug messages for reputation and Delve level-ups - [localization.lua]
 - Added - Reputation rank-up tracking for The Severed Threads sub-factions [core.lua]